feat: abstracting table pagination
