+++ Client/src/Pages/Infrastructure/index.jsx
@@ -13,27 +13,22 @@ import {
 	// TablePagination,
 	// Typography,
 } from "@mui/material";
-import { Heading } from "../../Components/Heading";
-
+import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
 import { useTheme } from "@emotion/react";
 import Greeting from "../../Utils/greeting";
-import Breadcrumbs from "../../Components/Breadcrumbs";
-import { StatusLabel } from "../../Components/Label";
-import { Gauge } from "../../Components/Charts/Gauge";
 import GearIcon from "../../Assets/icons/settings-bold.svg?react";
 import CPUChipIcon from "../../Assets/icons/cpu-chip.svg?react";
+import Breadcrumbs from "../../Components/Breadcrumbs";
+import { StatusLabel } from "../../Components/Label";
+import { Heading } from "../../Components/Heading";
+import { Gauge } from "../../Components/Charts/Gauge";
+import { getInfrastructureMonitorsByTeamId } from "../../Features/InfrastructureMonitors/infrastructureMonitorsSlice";
+import useUtils from "../Monitors/utils";
+import { Pagination } from "./components/TablePagination";
 
-const mockedData = {
-	ip: "https://192.168.1.30",
-	status: "up",
-	processor: "2Ghz",
-	cpu: 80,
-	mem: 50,
-	disk: 70,
-};
-
-const ROWS = Array.from(Array(20).keys()).map(() => mockedData);
-console.log(ROWS);
+// const ROWS = Array.from(Array(20).keys()).map(() => mockedData);
 
 const columns = [
 	{ label: "Host" },
@@ -69,8 +64,6 @@ Apply to Monitor Table, and Account/Team.
 Analyze existing BasicTable
 */
 
-import { useNavigate } from "react-router-dom";
-
 /**
  * This is the Infrastructure monitoring page. This is a work in progress
  *
@@ -78,9 +71,33 @@ import { useNavigate } from "react-router-dom";
  * @returns {JSX.Element} The infrastructure monitoring page.
  */
 
-function Infrastructure() {
+function Infrastructure(/* {isAdmin} */) {
 	const theme = useTheme();
+	const { determineState } = useUtils();
 	const navigate = useNavigate();
+	const dispatch = useDispatch();
+	const { authToken } = useSelector((state) => state.auth);
+	const { isLoading, monitorsSummary, msg, success, ...rest } = useSelector(
+		(state) => state.infrastructureMonitors
+	);
+
+	const { monitorCounts = {}, monitors = [] } = monitorsSummary;
+	const { total: totalMonitors = 0 } = monitorCounts;
+	console.log({ monitors });
+	const monitorsAsRows = monitors.map((monitor) => ({
+		ip: monitor.name,
+		status: determineState(monitor),
+		processor: "2Ghz" /* How to retrieve that?  */,
+		cpu: 80 /* How to retrieve that?  */,
+		mem: 50 /* How to retrieve that?  */,
+		disk: 70 /* How to retrieve that?  */,
+	}));
+
+	useEffect(() => {
+		dispatch(getInfrastructureMonitorsByTeamId(authToken));
+	}, [authToken]);
+	// console.log({ isLoading, monitorsSummary, msg, success, rest });
+
 	return (
 		<Stack
 			component="main"
@@ -134,7 +151,7 @@ function Infrastructure() {
 						borderColor={theme.palette.border.light}
 						backgroundColor={theme.palette.background.accent}
 					>
-						5
+						{totalMonitors}
 					</Box>
 				</Stack>
 				<TableContainer
@@ -170,142 +187,81 @@ function Infrastructure() {
 							</TableRow>
 						</TableHead>
 						<TableBody>
-							{ROWS.map((row, index) => (
-								<TableRow key={index}>
-									{/* TODO iterate over column and get column id, applying row[column.id] */}
-									<TableCell>{row.ip}</TableCell>
-									<TableCell align="center">
-										<StatusLabel
-											status={row.status}
-											text={row.status}
-											/* Use capitalize inside of Status Label */
-											/* Update component so we don't need to pass text and status separately*/
-											customStyles={{ textTransform: "capitalize" }}
-										/>
-									</TableCell>
-									<TableCell align="center">
-										<Stack
-											direction={"row"}
-											justifyContent={"center"}
-											alignItems={"center"}
-											gap=".25rem"
-										>
-											<CPUChipIcon
-												width={20}
-												height={20}
+							{
+								/* ROWS */ monitorsAsRows.map((row, index) => (
+									<TableRow key={index}>
+										{/* TODO iterate over column and get column id, applying row[column.id] */}
+										<TableCell>{row.ip}</TableCell>
+										<TableCell align="center">
+											<StatusLabel
+												status={row.status}
+												text={row.status}
+												/* Use capitalize inside of Status Label */
+												/* Update component so we don't need to pass text and status separately*/
+												customStyles={{ textTransform: "capitalize" }}
 											/>
-											{row.processor}
-										</Stack>
-									</TableCell>
-									<TableCell align="center">
-										<Gauge
-											progressValue={row.cpu}
-											containerWidth={60}
-										/>
-									</TableCell>
-									<TableCell align="center">
-										<Gauge
-											progressValue={row.mem}
-											containerWidth={60}
-										/>
-									</TableCell>
-									<TableCell align="center">
-										<Gauge
-											progressValue={row.disk}
-											containerWidth={60}
-										/>
-									</TableCell>
-									<TableCell align="center">
-										{/* Get ActionsMenu from Monitor Table and create a component */}
-										<IconButton
-											sx={{
-												"& svg path": {
-													stroke: theme.palette.other.icon,
-												},
-											}}
-										>
-											<GearIcon
-												width={20}
-												height={20}
+										</TableCell>
+										<TableCell align="center">
+											<Stack
+												direction={"row"}
+												justifyContent={"center"}
+												alignItems={"center"}
+												gap=".25rem"
+											>
+												<CPUChipIcon
+													width={20}
+													height={20}
+												/>
+												{row.processor}
+											</Stack>
+										</TableCell>
+										<TableCell align="center">
+											<Gauge
+												progressValue={row.cpu}
+												containerWidth={60}
 											/>
-										</IconButton>
-									</TableCell>
-								</TableRow>
-							))}
+										</TableCell>
+										<TableCell align="center">
+											<Gauge
+												progressValue={row.mem}
+												containerWidth={60}
+											/>
+										</TableCell>
+										<TableCell align="center">
+											<Gauge
+												progressValue={row.disk}
+												containerWidth={60}
+											/>
+										</TableCell>
+										<TableCell align="center">
+											{/* Get ActionsMenu from Monitor Table and create a component */}
+											<IconButton
+												sx={{
+													"& svg path": {
+														stroke: theme.palette.other.icon,
+													},
+												}}
+											>
+												<GearIcon
+													width={20}
+													height={20}
+												/>
+											</IconButton>
+										</TableCell>
+									</TableRow>
+								))
+							}
 						</TableBody>
 					</Table>
 				</TableContainer>
-				{/* <Stack
-					direction="row"
-					alignItems="center"
-					justifyContent="space-between"
-					marginTop={8}
-				>
-					<Typography
-						// px={theme.spacing(2)}
-						variant="body2"
-						// sx={{ opacity: 0.7 }}
-					>
-						Showing {getRange()} of {monitorCount} monitor(s)
-					</Typography>
-					<TablePagination
-						component="div"
-						count={monitorCount}
-						page={page}
-						onPageChange={handleChangePage}
-						rowsPerPage={rowsPerPage}
-						rowsPerPageOptions={[5, 10, 15, 25]}
-						onRowsPerPageChange={handleChangeRowsPerPage}
-						ActionsComponent={TablePaginationActions}
-						labelRowsPerPage="Rows per page"
-						labelDisplayedRows={({ page, count }) =>
-							`Page ${page + 1} of ${Math.max(0, Math.ceil(count / rowsPerPage))}`
-						}
-						slotProps={{
-							select: {
-								MenuProps: {
-									keepMounted: true,
-									disableScrollLock: true,
-									PaperProps: {
-										className: "pagination-dropdown",
-										sx: {
-											mt: 0,
-											mb: theme.spacing(2),
-										},
-									},
-									transformOrigin: { vertical: "bottom", horizontal: "left" },
-									anchorOrigin: { vertical: "top", horizontal: "left" },
-									sx: {
-										mt: theme.spacing(-2),
-									},
-								},
-								inputProps: { id: "pagination-dropdown" },
-								IconComponent: SelectorVertical,
-								sx: {
-									ml: theme.spacing(4),
-									mr: theme.spacing(12),
-									minWidth: theme.spacing(20),
-									textAlign: "left",
-									"&.Mui-focused > div": {
-										backgroundColor: theme.palette.background.main,
-									},
-								},
-							},
-						}}
-						sx={{
-							color: theme.palette.text.secondary,
-							"& svg path": {
-								stroke: theme.palette.text.tertiary,
-								strokeWidth: 1.3,
-							},
-							"& .MuiSelect-select": {
-								border: 1,
-								borderColor: theme.palette.border.light,
-								borderRadius: theme.shape.borderRadius,
-							},
-						}}
-					/>
-				</Stack> */}
+				{/* 
+				  TODO continue creating pagination component. It should change the current page, which will trigger refetch? 
+				
+				*/}
+				{/* <Pagination
+					// monitorCount={totalMonitors}
+					page={0}
+				/> */}
 			</Stack>
 		</Stack>
 	);
