+++ Client/src/Pages/Monitors/Home/MonitorTable/index.jsx
@@ -1,4 +1,15 @@
 import PropTypes from "prop-types";
+import { useState, useEffect, memo, useCallback, useRef } from "react";
+import { useNavigate } from "react-router";
+import { useDispatch, useSelector } from "react-redux";
+import { useTheme } from "@emotion/react";
+import useUtils from "../../utils";
+
+import { setRowsPerPage } from "../../../../Features/UI/uiSlice";
+import { logger } from "../../../../Utils/Logger";
+import { jwtDecode } from "jwt-decode";
+import { networkService } from "../../../../main";
+
 import {
 	TableContainer,
 	Table,
@@ -8,106 +19,18 @@ import {
 	TableBody,
 	Paper,
 	Box,
-	TablePagination,
-	Stack,
-	Typography,
-	Button,
 	CircularProgress,
 } from "@mui/material";
+import ActionsMenu from "../actionsMenu";
+import Host from "../host";
+import { StatusLabel } from "../../../../Components/Label";
+import { TableBodySkeleton } from "./Skeleton";
+import BarChart from "../../../../Components/Charts/BarChart";
+
 import ArrowDownwardRoundedIcon from "@mui/icons-material/ArrowDownwardRounded";
 import ArrowUpwardRoundedIcon from "@mui/icons-material/ArrowUpwardRounded";
 
-import { setRowsPerPage } from "../../../../Features/UI/uiSlice";
-import { useState, useEffect, memo, useCallback, useRef } from "react";
-import { useNavigate } from "react-router";
-import { logger } from "../../../../Utils/Logger";
-import Host from "../host";
-import { StatusLabel } from "../../../../Components/Label";
-import { jwtDecode } from "jwt-decode";
-import { useDispatch, useSelector } from "react-redux";
-import { networkService } from "../../../../main";
-import { useTheme } from "@emotion/react";
-import BarChart from "../../../../Components/Charts/BarChart";
-import LeftArrowDouble from "../../../../assets/icons/left-arrow-double.svg?react";
-import RightArrowDouble from "../../../../assets/icons/right-arrow-double.svg?react";
-import LeftArrow from "../../../../assets/icons/left-arrow.svg?react";
-import RightArrow from "../../../../assets/icons/right-arrow.svg?react";
-import SelectorVertical from "../../../../assets/icons/selector-vertical.svg?react";
-import ActionsMenu from "../actionsMenu";
-import useUtils from "../../utils";
-import { TableBodySkeleton } from "./Skeleton";
-
-/**
- * Component for pagination actions (first, previous, next, last).
- *
- * @component
- * @param {Object} props
- * @param {number} props.count - Total number of items.
- * @param {number} props.page - Current page number.
- * @param {number} props.rowsPerPage - Number of rows per page.
- * @param {function} props.onPageChange - Callback function to handle page change.
- *
- * @returns {JSX.Element} Pagination actions component.
- */
-const TablePaginationActions = (props) => {
-	const { count, page, rowsPerPage, onPageChange } = props;
-	const handleFirstPageButtonClick = (event) => {
-		onPageChange(event, 0);
-	};
-	const handleBackButtonClick = (event) => {
-		onPageChange(event, page - 1);
-	};
-	const handleNextButtonClick = (event) => {
-		onPageChange(event, page + 1);
-	};
-	const handleLastPageButtonClick = (event) => {
-		onPageChange(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-	};
-
-	return (
-		<Box sx={{ flexShrink: 0, ml: "24px" }}>
-			<Button
-				variant="group"
-				onClick={handleFirstPageButtonClick}
-				disabled={page === 0}
-				aria-label="first page"
-			>
-				<LeftArrowDouble />
-			</Button>
-			<Button
-				variant="group"
-				onClick={handleBackButtonClick}
-				disabled={page === 0}
-				aria-label="previous page"
-			>
-				<LeftArrow />
-			</Button>
-			<Button
-				variant="group"
-				onClick={handleNextButtonClick}
-				disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-				aria-label="next page"
-			>
-				<RightArrow />
-			</Button>
-			<Button
-				variant="group"
-				onClick={handleLastPageButtonClick}
-				disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-				aria-label="last page"
-			>
-				<RightArrowDouble />
-			</Button>
-		</Box>
-	);
-};
-
-TablePaginationActions.propTypes = {
-	count: PropTypes.number.isRequired,
-	page: PropTypes.number.isRequired,
-	rowsPerPage: PropTypes.number.isRequired,
-	onPageChange: PropTypes.func.isRequired,
-};
+import { Pagination } from "../../../Infrastructure/components/TablePagination";
 
 const MonitorTable = ({ isAdmin, filter, setIsSearching, isSearching }) => {
 	const theme = useTheme();
@@ -117,6 +40,7 @@ const MonitorTable = ({ isAdmin, filter, setIsSearching, isSearching }) => {
 
 	const { rowsPerPage } = useSelector((state) => state.ui.monitors);
 	const authState = useSelector((state) => state.auth);
+
 	const [page, setPage] = useState(0);
 	const [monitors, setMonitors] = useState([]);
 	const [monitorCount, setMonitorCount] = useState(0);
@@ -195,11 +119,11 @@ const MonitorTable = ({ isAdmin, filter, setIsSearching, isSearching }) => {
 	 * Helper function to calculate the range of displayed rows.
 	 * @returns {string}
 	 */
-	const getRange = () => {
-		let start = page * rowsPerPage + 1;
-		let end = Math.min(page * rowsPerPage + rowsPerPage, monitorCount);
-		return `${start} - ${end}`;
-	};
+	// const getRange = () => {
+	// 	let start = page * rowsPerPage + 1;
+	// 	let end = Math.min(page * rowsPerPage + rowsPerPage, monitorCount);
+	// 	return `${start} - ${end}`;
+	// };
 
 	const handleSort = async (field) => {
 		let order = "";
@@ -392,78 +316,13 @@ const MonitorTable = ({ isAdmin, filter, setIsSearching, isSearching }) => {
 					</TableBody>
 				</Table>
 			</TableContainer>
-			<Stack
-				direction="row"
-				alignItems="center"
-				justifyContent="space-between"
-				px={theme.spacing(4)}
-				marginTop={8}
-			>
-				<Typography
-					px={theme.spacing(2)}
-					variant="body2"
-					sx={{ opacity: 0.7 }}
-				>
-					Showing {getRange()} of {monitorCount} monitor(s)
-				</Typography>
-				<TablePagination
-					component="div"
-					count={monitorCount}
-					page={page}
-					onPageChange={handleChangePage}
-					rowsPerPage={rowsPerPage}
-					rowsPerPageOptions={[5, 10, 15, 25]}
-					onRowsPerPageChange={handleChangeRowsPerPage}
-					ActionsComponent={TablePaginationActions}
-					labelRowsPerPage="Rows per page"
-					labelDisplayedRows={({ page, count }) =>
-						`Page ${page + 1} of ${Math.max(0, Math.ceil(count / rowsPerPage))}`
-					}
-					slotProps={{
-						select: {
-							MenuProps: {
-								keepMounted: true,
-								disableScrollLock: true,
-								PaperProps: {
-									className: "pagination-dropdown",
-									sx: {
-										mt: 0,
-										mb: theme.spacing(2),
-									},
-								},
-								transformOrigin: { vertical: "bottom", horizontal: "left" },
-								anchorOrigin: { vertical: "top", horizontal: "left" },
-								sx: {
-									mt: theme.spacing(-2),
-								},
-							},
-							inputProps: { id: "pagination-dropdown" },
-							IconComponent: SelectorVertical,
-							sx: {
-								ml: theme.spacing(4),
-								mr: theme.spacing(12),
-								minWidth: theme.spacing(20),
-								textAlign: "left",
-								"&.Mui-focused > div": {
-									backgroundColor: theme.palette.background.main,
-								},
-							},
-						},
-					}}
-					sx={{
-						color: theme.palette.text.secondary,
-						"& svg path": {
-							stroke: theme.palette.text.tertiary,
-							strokeWidth: 1.3,
-						},
-						"& .MuiSelect-select": {
-							border: 1,
-							borderColor: theme.palette.border.light,
-							borderRadius: theme.shape.borderRadius,
-						},
-					}}
-				/>
-			</Stack>
+			<Pagination
+				page={page}
+				rowsPerPage={rowsPerPage}
+				handleChangePage={handleChangePage}
+				handleChangeRowsPerPage={handleChangeRowsPerPage}
+				monitorCount={monitorCount}
+			/>
 		</Box>
 	);
 };
